+++ tests/testthat/test_download_terms.R
@@ -1,0 +1,3 @@
+test_that("Error when incorrect species", {
+  expect_error(download_functional_terms("not a species"))
+})