+++ NEWS.md
@@ -6,3 +6,11 @@
  
   - A lot of minor tweaks
   - Documentation updates
+
+## Version 0.1.2
+
+ - Added functions `list_species()` and `download_feature_information()`.
+ - Split function `download_functional_terms()` into `download_functional_terms()` and `prepare_functional_terms()` to allow for more flexibility with feature identifiers.
+ - Wrote new section of the vignette with an advanced example.
+ - Added mouse RNA-seq data for the example.
+ - Added Ensembl settings for all species in the `species.json` file.